feat!: update to libp2p v3 api
Refactors code to conform to the `libp2p@3.x.x` API.

- Encrypted streams are now EventTargets

BREAKING CHANGE: Must be used with `libp2p@3.x.x`, it cannot be used with earlier versions
