+++ benchmarks/benchmarks.js
@@ -1,5 +1,3 @@
-'use strict'
-
 const Benchmark = require('benchmark')
 const utils = require('../test/utils')
 const suite = new Benchmark.Suite('gossipsub')
