+++ package.json
@@ -73,14 +73,15 @@
   },
   "homepage": "https://github.com/ChainSafe/js-libp2p-gossipsub#readme",
   "dependencies": {
-    "@libp2p/crypto": "^5.0.0",
-    "@libp2p/interface": "^2.0.0",
-    "@libp2p/interface-internal": "^2.0.0",
-    "@libp2p/peer-id": "^5.0.0",
+    "@libp2p/crypto": "^5.1.9",
+    "@libp2p/interface": "^3.0.0",
+    "@libp2p/interface-internal": "^3.0.0",
+    "@libp2p/peer-id": "^6.0.0",
     "@libp2p/pubsub": "^10.0.0",
-    "@multiformats/multiaddr": "^12.1.14",
+    "@libp2p/utils": "^7.0.0",
+    "@multiformats/multiaddr": "^13.0.1",
     "denque": "^2.1.0",
-    "it-length-prefixed": "^9.0.4",
+    "it-length-prefixed": "^10.0.1",
     "it-pipe": "^3.0.1",
     "it-pushable": "^3.2.3",
     "multiformats": "^13.0.1",
@@ -89,38 +90,30 @@
     "uint8arrays": "^5.0.1"
   },
   "devDependencies": {
-    "@chainsafe/as-sha256": "^0.4.1",
-    "@dapplion/benchmark": "^0.2.4",
-    "@libp2p/floodsub": "^10.0.0",
-    "@libp2p/interface-compliance-tests": "^6.0.0",
-    "@libp2p/logger": "^5.0.0",
-    "@libp2p/peer-store": "^11.0.0",
-    "@types/node": "^20.11.6",
+    "@chainsafe/as-sha256": "^1.2.0",
+    "@dapplion/benchmark": "^1.0.0",
+    "@libp2p/floodsub": "^11.0.0",
+    "@libp2p/logger": "^6.0.0",
+    "@libp2p/peer-store": "^12.0.0",
+    "@types/node": "^22.18.1",
     "@types/sinon": "^17.0.3",
     "abortable-iterator": "^5.1.0",
-    "aegir": "^44.1.1",
+    "aegir": "^47.0.21",
     "datastore-core": "^10.0.0",
     "delay": "^6.0.0",
     "it-all": "^3.0.6",
     "mkdirp": "^3.0.1",
     "p-defer": "^4.0.0",
-    "p-event": "^6.0.0",
-    "p-retry": "^6.2.0",
-    "p-wait-for": "^5.0.2",
+    "p-event": "^7.0.0",
+    "p-retry": "^7.0.0",
+    "p-wait-for": "^6.0.0",
     "protons": "^7.5.0",
-    "sinon": "^18.0.1",
+    "sinon": "^21.0.0",
     "sinon-ts": "^2.0.0",
     "time-cache": "^0.3.0"
   },
   "engines": {
     "npm": ">=8.7.0"
-  },
-  "eslintConfig": {
-    "extends": "ipfs",
-    "ignorePatterns": [
-      "src/message/rpc*",
-      "!.aegir.js"
-    ]
   },
   "contributors": [
     "Cayman <caymannava@gmail.com>",
