+++ src/types.ts
@@ -1,5 +1,7 @@
-import { type Message, TopicValidatorResult, type PrivateKey, type PeerId } from '@libp2p/interface'
+import { TopicValidatorResult } from './index.ts'
+import type { Message } from './index.ts'
 import type { RPC } from './message/rpc.js'
+import type { PrivateKey, PeerId } from '@libp2p/interface'
 import type { Multiaddr } from '@multiformats/multiaddr'
 
 export type MsgIdStr = string
