+++ src/utils/publishConfig.ts
@@ -1,6 +1,7 @@
 import { publicKeyToProtobuf } from '@libp2p/crypto/keys'
-import { StrictSign, StrictNoSign } from '@libp2p/interface'
-import { type PublishConfig, PublishConfigType } from '../types.js'
+import { StrictSign, StrictNoSign } from '../index.ts'
+import { PublishConfigType } from '../types.js'
+import type { PublishConfig } from '../types.js'
 import type { PeerId, PrivateKey } from '@libp2p/interface'
 
 /**
