+++ test/benchmark/asyncIterable.test.ts
@@ -3,14 +3,12 @@ import { abortableSource } from 'abortable-iterator'
 import all from 'it-all'
 import { pipe } from 'it-pipe'
 
-/* eslint-disable generator-star-spacing */
-
 describe('abortableSource cost', function () {
   const n = 10000
   const bytes = new Uint8Array(200)
   const controller = new AbortController()
 
-  async function* bytesSource (): AsyncGenerator<Uint8Array, void, unknown> {
+  async function * bytesSource (): AsyncGenerator<Uint8Array, void, unknown> {
     let i = 0
     while (i++ < n) {
       yield bytes
@@ -40,14 +38,14 @@ describe('abortableSource cost', function () {
 describe('pipe extra iterables cost', function () {
   const n = 10000
 
-  async function* numberSource (): AsyncGenerator<number, void, unknown> {
+  async function * numberSource (): AsyncGenerator<number, void, unknown> {
     let i = 0
     while (i < n) {
       yield i++
     }
   }
 
-  async function* numberTransform (source: AsyncIterable<number>): AsyncIterable<number> {
+  async function * numberTransform (source: AsyncIterable<number>): AsyncIterable<number> {
     for await (const num of source) {
       yield num + 1
     }
