+++ test/signature-policy.spec.ts
@@ -1,14 +1,14 @@
 import { stop } from '@libp2p/interface'
-import { mockNetwork } from '@libp2p/interface-compliance-tests/mocks'
 import { expect } from 'aegir/chai'
 import { pEvent } from 'p-event'
 import {
   connectAllPubSubNodes,
   connectPubsubNodes,
   createComponents,
-  createComponentsArray,
-  type GossipSubAndComponents
+  createComponentsArray
+
 } from './utils/create-pubsub.js'
+import type { GossipSubAndComponents } from './utils/create-pubsub.js'
 
 describe('signature policy', () => {
   describe('strict-sign', () => {
@@ -16,7 +16,6 @@ describe('signature policy', () => {
     let nodes: GossipSubAndComponents[]
 
     beforeEach(async () => {
-      mockNetwork.reset()
       nodes = await createComponentsArray({
         number: numNodes,
         connected: false,
@@ -32,7 +31,6 @@ describe('signature policy', () => {
 
     afterEach(async () => {
       await stop(...nodes.reduce<any[]>((acc, curr) => acc.concat(curr.pubsub, ...Object.entries(curr.components)), []))
-      mockNetwork.reset()
     })
 
     it('should publish a message', async () => {
@@ -116,7 +114,6 @@ describe('signature policy', () => {
     let nodes: GossipSubAndComponents[]
 
     beforeEach(async () => {
-      mockNetwork.reset()
       nodes = await createComponentsArray({
         number: numNodes,
         connected: false,
@@ -132,7 +129,6 @@ describe('signature policy', () => {
 
     afterEach(async () => {
       await stop(...nodes.reduce<any[]>((acc, curr) => acc.concat(curr.pubsub, ...Object.entries(curr.components)), []))
-      mockNetwork.reset()
     })
 
     it('should publish a message', async () => {
